# docs/swagger.yaml
basePath: /
definitions:
  doc.CPDetail:
    properties:
      keterangan:
        type: string
      name:
        type: string
      ok:
        type: boolean
    type: object
  doc.DeviceDocForm:
    properties:
      checkpoint:
        items:
          $ref: '#/definitions/doc.CPDetail'
        type: array
      device_id:
        type: string
    type: object
  doc.UpdateDeviceDocForm:
    properties:
      checkpoint:
        items:
          $ref: '#/definitions/doc.CPDetail'
        type: array
    type: object
...
  handler.toaForm:
    properties:
      kode:
        type: string
      lokasi:
        type: string
      nama:
        type: string
      posisi:
        type: string
    type: object
  handler.updateCheckpointForm:
    properties:
      checkpoint:
        items:
          type: string
        type: array
    type: object
  handler.upsForm:
    properties:
      departemen:
        type: string
      lokasi:
        type: string
      nama:
        type: string
      no_seri:
        type: string
      tipe:
        type: string
    type: object
...
info:
  contact: {}
  description: Sistem Pencatatan Maintenance IT Backend API
  title: Sistem Pencatatan Maintenance IT Backend
paths:
...
  /api/checkpoint/toa:
    get:
      operationId: get-toa-checkpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get toa checkpoint
      tags:
      - Checkpoint
    put:
      consumes:
      - application/json
      operationId: update-toa-checkpoint
      parameters:
      - description: Checkpoint
        in: body
        name: checkpoint
        required: true
        schema:
          $ref: '#/definitions/handler.updateCheckpointForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update toa checkpoint
      tags:
      - Checkpoint
...
  /api/doc/toa:
    post:
      operationId: create-toa-document
      parameters:
      - description: TOA Document Form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/doc.DeviceDocForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Create toa document
      tags:
      - Doc TOA
  /api/doc/toa/{id}:
    delete:
      operationId: delete-toa-document-by-id
      parameters:
      - description: TOA Document ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Delete toa document by ID
      tags:
      - Doc TOA
    get:
      operationId: get-toa-document-by-id
      parameters:
      - description: TOA Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get toa document by ID
      tags:
      - Doc TOA
    put:
      operationId: update-toa-document-by-id
      parameters:
      - description: TOA Document ID
        in: path
        name: id
        required: true
        type: string
      - description: TOA Document Form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/doc.UpdateDeviceDocForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update toa document by ID
      tags:
      - Doc TOA
  /api/doc/toas:
    get:
      operationId: get-all-toas-documents
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get all toas documents
      tags:
      - Doc TOA
  /api/doc/ups:
    get:
      operationId: get-all-ups-documents
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get all ups documents
      tags:
      - Doc UPS
    post:
      operationId: create-new-ups-document
      parameters:
      - description: UPS Document Form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/doc.DeviceDocForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Create new ups document
      tags:
      - Doc UPS
  /api/doc/ups/{id}:
    delete:
      operationId: delete-ups-document-by-id
      parameters:
      - description: UPS Document ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Delete ups document by ID
      tags:
      - Doc UPS
    get:
      operationId: get-ups-document-by-id
      parameters:
      - description: UPS Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get ups document by ID
      tags:
      - Doc UPS
    put:
      operationId: update-ups-document-by-id
      parameters:
      - description: UPS Document ID
        in: path
        name: id
        required: true
        type: string
      - description: UPS Document Form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/doc.UpdateDeviceDocForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update ups document by ID
      tags:
      - Doc UPS
...
  /api/toa:
    post:
      operationId: create-new-toa
      parameters:
      - description: TOA Form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.toaForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Create new toa
      tags:
      - Device TOA
  /api/toa/{id}:
    delete:
      operationId: delete-toa-by-id
      parameters:
      - description: TOA ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Delete toa by id
      tags:
      - Device TOA
    get:
      operationId: get-toa-by-id
      parameters:
      - description: TOA ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get toa by id
      tags:
      - Device TOA
    put:
      operationId: update-toa-by-id
      parameters:
      - description: TOA ID
        in: path
        name: id
        required: true
        type: string
      - description: TOA Form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.toaForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update toa by id
      tags:
      - Device TOA
  /api/toas:
    get:
      operationId: get-all-toas
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get all toas
      tags:
      - Device TOA
  /api/ups:
    get:
      operationId: get-all-ups
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get all ups
      tags:
      - Device UPS
    post:
      operationId: create-new-ups
      parameters:
      - description: UPS Form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.upsForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Create new ups
      tags:
      - Device UPS
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"